Move actual functionality into a lib
